perf(portfolio): use ``use cache'` directive
