+++ apps/portfolio/components/section/index.ts
@@ -1,0 +1,2 @@
+export * from './section'
+export { default } from './section'
