+++ apps/portfolio/package.json
@@ -5,7 +5,6 @@
     "@portabletext/react": "^4.0.0",
     "@sanity/client": "^7.0.0",
     "@vercel/analytics": "^1.4.1",
-    "clsx": "^2.1.1",
     "next": "16.0.0-canary.1",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
@@ -51,8 +50,8 @@
     "url": "https://github.com/ykzts/ykzts.git"
   },
   "scripts": {
-    "build": "next build --turbopack",
-    "dev": "next --turbopack",
+    "build": "next build",
+    "dev": "next",
     "lighthouse": "lhci autorun",
     "start": "next start",
     "test": "vitest --run",
