+++ apps/studio/package.json
@@ -30,8 +30,8 @@
     "url": "https://github.com/ykzts/ykzts.git"
   },
   "scripts": {
-    "build": "next build --turbopack",
-    "dev": "next -p 10000 --turbopack",
+    "build": "next build",
+    "dev": "next -p 10000",
     "start": "next start -p 10000",
     "test": "echo \"No tests configured for studio app\"",
     "typegen": "next typegen"
